implements deferred rebel task configuration - only when JavaPlugin applied
The rebel task makes sense only when the project is Java or derivate.
By default, a dummy task is added that fails the build saying there
is no JavaPlugin applied. When the build applies the JavaPlugin (or
any other plugin that applies it, like GroovyPlugin, ScalaPlugin, ...)
the dummy task gets replaced with the correct one. This makes for a
nice sanity check that rebel is only configured for Java projects.

Appropriate tests also added.