JR-11052 Change plugin to create a xml generation task per SourceSet on gradle4+ Do not generate dummy task throwing exception if "java" plugin not configured
Makes it easier to match remote id with IDE generated ones and expands the use of rebel.xml to integration tests and other uses of SourceSet.

Not creating dummy task allows plugin to be applied on all subprojects without fear that a single module without "java" plugin causes "generateRebel" call in root to fail.